kernel: Introduce panic logstore
Well, this can be an alternative of pstore but you shouldn't use this if your
pstore works fine. It's not stable and doesn't give as many information as
pstore does.

It optimistically assumes that filesystems and related storage drivers are not
to blame for this panic, otherwise the behavior can be unpredictable. Also, it
can't catch logs before the target partition is properly mounted.

The path of the log file is defined in the LOG_FILE_PATH macro.

Test: echo 1 > /sys/module/panic_logstore/parameters/trigger

Signed-off-by: LibXZR <i@xzr.moe>